Created validatePuzzle function and fixed issue with unit tests.  (#14)
* Changelog:
Created "validatePuzzle" function that handles all puzzle validation. Includes validation that determines if there are duplicate values for rows, columns, or box. Fixed issue with tests that are expecting errors to be thrown. If no error was thrown the test passed because it was in a try-catch block. Fixed so that if test incorrectly throws no errors it throws an error called "NoErrorThrownError".

* Removing unneeded code.

* Cleaning up comments.

* Added static getBoxStart functions

* Changed how checking for duplicates in boxes

* Changed row and column validation to same as box

* Added more Duplicate Row and Column Tests.

* Added Duplicate Box tests

Co-authored-by: Gregory <gregorytravis711@gmail.com>