Searched groups (#50)
Currently for every step every group is checked for every strategy until an applicable one is found. The problem with this is that most groups don’t get changed each step resulting in searching unchanged groups for the same strategies step after step. This PR seeks to cache which groups have been checked already for a strategy for as long as the group remains unchanged and use that cache to avoid repeating work thus boosting performance substantially.