Lessons upgrade (#63)
Switched lessons API to using S3 so lessons can be modified and new lessons can be added without having to update the package with new hardcoded values. Updated README with how to update the content in S3. Also added Lessons.getTutorial() to get the first few lessons in one in order to onboard new users (and updated README with its usage).